feat(indexer): add SSE event catch-up/replay support to transaction events stream (#1948)
## Description

Adds catch-up/replay support to the indexer's
`/transactions/events/stream` SSE endpoint. Previously, clients that
disconnected would permanently miss events. Now clients can reconnect
and recover all missed events.

## Motivation

SSE clients (e.g. the wallet SDK event watcher) that experience network
interruptions had no recovery path — missed events were gone. This is a
significant reliability gap for anything that depends on event-driven
state updates.

## Changes

- **`TransactionEvent`**: added `id: i64` field populated from the DB
auto-increment primary key
- **Broadcast order**: reversed in the worker — DB insert happens before
broadcast so events always have stable IDs when clients receive them
- **`get_events_after_id`**: new ascending cursor-based DB query for
replaying missed events
- **`StreamTransactionEventsRequest`**: added `after_id: Option<i64>`
query parameter; also accepts standard `Last-Event-ID` header
- **SSE handler rewrite**: subscribe to broadcast channel first, replay
from DB, dedup by ID, then drain live channel — ensures no gaps
- **Replay cap**: 10,000 events maximum to prevent DoS from clients
requesting replay from genesis
- **TypeScript bindings**: updated to reflect new fields
- **Version bumps**: `tari_indexer_client` 0.27.1→0.27.2, `ootle-rs`
0.3.0→0.3.1, `ootle-ts-bindings` 1.30.1→1.30.2, `indexer-client`
1.3.1→1.3.2
# Conflicts:
#	Cargo.lock
#	Cargo.toml

+++ applications/tari_indexer/src/storage_sqlite/schema.rs
@@ -13,7 +13,7 @@ diesel::table! {
 
 diesel::table! {
     events (id) {
-        id -> Integer,
+        id -> BigInt,
         template_address -> Text,
         tx_hash -> Text,
         topic -> Text,
